requirements.txt
asposestorage==1.0.2
keras==2.8.0
librosa==0.9.1
matplotlib==3.5.0
numpy==1.21.2
pandas==1.3.2
Pillow==9.0.1
PySide2==5.15.2.1
scikit_learn==1.0.2
scipy==1.7.2
seaborn==0.11.2
SoundFile==0.10.3.post1
XlsxWriter==3.0.3
